fix: floating point precision error
For the integration methods we use x = x + h to get the correct x of each iteration. But as we are summing floating point numbers, over large n the floating point precision error adds up. It is insignificant for small n though. Using x = a + i * h will ensure we always get the correct value of x.